# feat: add cuda sampling kernels for top_k and top_p