--- a/aphrodite/_custom_ops.py
+++ b/aphrodite/_custom_ops.py
@@ -1,6 +1,6 @@
 import contextlib
 import functools
-from typing import List, Optional, Tuple, Type
+from typing import List, Optional, Tuple, Type, Union
 
 import torch
 from loguru import logger
@@ -630,6 +630,122 @@ def get_graph_buffer_ipc_meta(fa: int) -> Tuple[List[str], List[int]]:
 def register_graph_buffers(fa: int, handles: List[str],
                            offsets: List[List[int]]) -> None:
     torch.ops._C_custom_ar.register_graph_buffers(fa, handles, offsets)
+
+
+# Sampling Kernels
+def sampling_from_probs(probs: torch.Tensor,
+                        uniform_samplers: torch.Tensor,
+                        deterministic: bool = True,
+                        check_nan: bool = False) -> torch.Tensor:
+    if check_nan and torch.any(torch.isnan(probs)):
+        raise ValueError("NaN detected in probs")
+    return torch.ops._C.sampling_from_probs(probs, uniform_samplers,
+                                            deterministic)
+
+def _to_tensor_scalar_tuple(x):
+    if isinstance(x, torch.Tensor):
+        return (x, 0)
+    else:
+        return (None, x)
+def top_p_sampling_from_probs(
+        probs: torch.Tensor,
+        uniform_samples: torch.Tensor,
+        top_p: Union[torch.Tensor, float],
+        deterministic: bool = True,
+        check_nan: bool = False) -> Tuple[torch.Tensor, torch.Tensor]:
+    if check_nan and torch.any(torch.isnan(probs)):
+        raise ValueError("NaN detected in probs")
+    return torch.ops._C.top_p_sampling_from_probs(
+        probs, uniform_samples, *_to_tensor_scalar_tuple(top_p), deterministic)
+
+def top_k_sampling_from_probs(
+        probs: torch.Tensor,
+        uniform_samples: torch.Tensor,
+        top_k: Union[torch.Tensor, int],
+        deterministic: bool = True,
+        check_nan: bool = False) -> Tuple[torch.Tensor, torch.Tensor]:
+    if check_nan and torch.any(torch.isnan(probs)):
+        raise ValueError("NaN detected in probs")
+    return torch.ops._C.top_k_sampling_from_probs(
+        probs, uniform_samples, *_to_tensor_scalar_tuple(top_k), deterministic)
+
+def min_p_sampling_from_probs(
+        probs: torch.Tensor,
+        uniform_samples: torch.Tensor,
+        min_p: Union[torch.Tensor, float],
+        deterministic: bool = True,
+        check_nan: bool = False) -> Tuple[torch.Tensor, torch.Tensor]:
+    if check_nan and torch.any(torch.isnan(probs)):
+        raise ValueError("NaN detected in probs")
+    return torch.ops._C.min_p_sampling_from_probs(
+        probs, uniform_samples, *_to_tensor_scalar_tuple(min_p), deterministic)
+
+def top_k_mask_logits(
+    logits: torch.Tensor,
+    top_k: Union[torch.Tensor, int],
+) -> torch.Tensor:
+    return torch.ops._C.top_k_mask_logits(logits,
+                                          *_to_tensor_scalar_tuple(top_k))
+
+def top_p_renorm_prob(
+    probs: torch.Tensor,
+    top_p: Union[torch.Tensor, float],
+) -> torch.Tensor:
+    return torch.ops._C.top_p_renorm_prob(probs,
+                                          *_to_tensor_scalar_tuple(top_p))
+
+def top_k_renorm_prob(
+    probs: torch.Tensor,
+    top_k: Union[torch.Tensor, int],
+) -> torch.Tensor:
+    return torch.ops._C.top_k_renorm_prob(probs,
+                                          *_to_tensor_scalar_tuple(top_k))
+
+def top_k_top_p_sampling_from_logits(
+    probs: torch.Tensor,
+    uniform_samples: torch.Tensor,
+    top_k: Union[torch.Tensor, int],
+    top_p: Union[torch.Tensor, float],
+    filter_apply_order: str = "top_k_first",
+    deterministic: bool = True,
+    check_nan: bool = False,
+) -> Tuple[torch.Tensor, torch.Tensor]:
+    if filter_apply_order == "top_k_first":
+        masked_logits = top_k_mask_logits(probs, top_k)
+        probs = torch.softmax(masked_logits, dim=-1)
+        return top_p_sampling_from_probs(probs, uniform_samples, top_p,
+                                         deterministic, check_nan)
+    elif filter_apply_order == "joint":
+        probs = torch.softmax(probs, dim=-1)
+        if check_nan and torch.any(torch.isnan(probs)):
+            raise ValueError("NaN detected in probs")
+        return torch.ops._C.top_k_top_p_sampling_from_logits(
+            probs, uniform_samples, *_to_tensor_scalar_tuple(top_k),
+            *_to_tensor_scalar_tuple(top_p), deterministic)
+    else:
+        raise ValueError(f"Invalid filter_apply_order: {filter_apply_order}")
+
+def top_k_top_p_sampling_from_probs(
+    probs: torch.Tensor,
+    uniform_samples: torch.Tensor,
+    top_k: Union[torch.Tensor, int],
+    top_p: Union[torch.Tensor, float],
+    filter_apply_order: str = "top_k_first",
+    deterministic: bool = True,
+    check_nan: bool = False,
+) -> Tuple[torch.Tensor, torch.Tensor]:
+    if filter_apply_order == "top_k_first":
+        renorm_probs = top_k_renorm_prob(probs, top_k)
+        return top_p_sampling_from_probs(renorm_probs, uniform_samples, top_p,
+                                         deterministic, check_nan)
+    elif filter_apply_order == "joint":
+        if check_nan and torch.any(torch.isnan(probs)):
+            raise ValueError("NaN detected in probs")
+        return torch.ops._C.top_k_top_p_sampling_from_probs(
+            probs, uniform_samples, *_to_tensor_scalar_tuple(top_k),
+            *_to_tensor_scalar_tuple(top_p), deterministic)
+    else:
+        raise ValueError(f"Invalid filter_apply_order: {filter_apply_order}")
 
 
 # TODO: remove this later